refactor(strategy-service): introduce event publishing abstraction

+++ services/strategy-service/src/main/java/com/harsha/strategy_service/domain/model/SymbolStatisticsState.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 @Setter
 public class SymbolStatisticsState {
     private final String symbol;
-    private long sampleCount;
+    private int sampleCount;
     private double meanReturn;
     private double returnVariance;
     private double meanVolatility;
